requirements.txt
torchaudio
sounddevice
taglib-python